+++ src/orchestrator/contracts/orchestrator-component-descriptor.js
@@ -1,0 +1,9 @@
+export class OrchestratorComponentDescriptor {
+  constructor({ componentName, responsibility, ownsImplementation = false }) {
+    this.componentName = componentName;
+    this.responsibility = responsibility;
+    this.ownsImplementation = Boolean(ownsImplementation);
+
+    Object.freeze(this);
+  }
+}